feat: move trace manager to other file
