+++ include/lifter/SimpleTraceManager.hpp
@@ -1,0 +1,25 @@
+#pragma once
+
+#include <map>
+#include <remill/BC/TraceLifter.h>
+
+using Memory = std::map<uint64_t, uint8_t>;
+
+class SimpleTraceManager : public remill::TraceManager {
+  public:
+    SimpleTraceManager(std::vector<X86Function> funcs);
+
+  protected:
+    void SetLiftedTraceDefinition(uint64_t addr,
+                                  llvm::Function *lifted_func) override;
+
+    llvm::Function *GetLiftedTraceDeclaration(uint64_t addr) override;
+
+    llvm::Function *GetLiftedTraceDefinition(uint64_t addr) override;
+
+    bool TryReadExecutableByte(uint64_t addr, uint8_t *byte) override;
+
+  public:
+    Memory memory;
+    std::unordered_map<uint64_t, llvm::Function *> traces;
+};
